==> src/ai/development/ML/2. data_preparation.py <==



# Creazione di due nuove colonne "temp diff" e "humidity diff", che rappresentano rispettivamente
# la differenza tra la temperatura esterna e interna all'arnia, e
# la differenza tra l'umidità interna ed esterna all'arnia
dataset["temp diff"] = dataset["hive temp"] - dataset["weather temp"]
dataset["humidity diff"] = dataset["hive humidity"] - dataset["weather humidity"]
